[datadog exporter] Move ProcessMetrics test to metrics package (#1488)
* Move test as per CR comments

* Fix getting cached value

* Remove unneeded dependencies and config sections
